+++ mobile/library/java/org/chromium/net/impl/NativeCronetEngineBuilderImpl.java
@@ -61,6 +61,7 @@ public class NativeCronetEngineBuilderImpl extends CronetEngineBuilderImpl {
   private String mAppId = "unspecified";
   private TrustChainVerification mTrustChainVerification = VERIFY_TRUST_CHAIN;
   private String mVirtualClusters = "[]";
+  private boolean mEnablePlatformCertificatesValidation = true;
 
   /**
    * Builder for Native Cronet Engine. Default config enables SPDY, disables QUIC and HTTP cache.
@@ -114,6 +115,7 @@ public class NativeCronetEngineBuilderImpl extends CronetEngineBuilderImpl {
         mH2ConnectionKeepaliveTimeoutSeconds, mH2ExtendKeepaliveTimeout, mMaxConnectionsPerHost,
         mStatsFlushSeconds, mStreamIdleTimeoutSeconds, mPerTryIdleTimeoutSeconds, mAppVersion,
         mAppId, mTrustChainVerification, mVirtualClusters, nativeFilterChain, platformFilterChain,
-        stringAccessors, keyValueStores, statSinks, mEnableSkipDNSLookupForProxiedRequests);
+        stringAccessors, keyValueStores, statSinks, mEnableSkipDNSLookupForProxiedRequests,
+        mEnablePlatformCertificatesValidation);
   }
 }