+++ mobile/test/cc/unit/envoy_config_test.cc
@@ -40,21 +40,21 @@ TEST(TestConfig, ConfigIsApplied) {
       .setDeviceOs("probably-ubuntu-on-CI");
   std::string config_str = engine_builder.generateConfigStr();
 
-  std::vector<std::string> must_contain = {
-      "- &stats_domain asdf.fake.website",
-      "- &connect_timeout 123s",
-      "- &dns_refresh_rate 456s",
-      "- &dns_fail_base_interval 789s",
-      "- &dns_fail_max_interval 987s",
-      "- &dns_query_timeout 321s",
-      "- &dns_preresolve_hostnames [hostname]",
-      "- &h2_connection_keepalive_idle_interval 0.222s",
-      "- &h2_connection_keepalive_timeout 333s",
-      "- &stats_flush_interval 654s",
-      "- &virtual_clusters [virtual-clusters]",
-      ("- &metadata { device_os: probably-ubuntu-on-CI, "
-       "app_version: 1.2.3, app_id: 1234-1234-1234 }"),
-  };
+  std::vector<std::string> must_contain = {"- &stats_domain asdf.fake.website",
+                                           "- &connect_timeout 123s",
+                                           "- &dns_refresh_rate 456s",
+                                           "- &dns_fail_base_interval 789s",
+                                           "- &dns_fail_max_interval 987s",
+                                           "- &dns_query_timeout 321s",
+                                           "- &dns_preresolve_hostnames [hostname]",
+                                           "- &h2_connection_keepalive_idle_interval 0.222s",
+                                           "- &h2_connection_keepalive_timeout 333s",
+                                           "- &stats_flush_interval 654s",
+                                           "- &virtual_clusters [virtual-clusters]",
+                                           ("- &metadata { device_os: probably-ubuntu-on-CI, "
+                                            "app_version: 1.2.3, app_id: 1234-1234-1234 }"),
+                                           R"(- &validation_context
+  trusted_ca:)"};
   for (const auto& string : must_contain) {
     ASSERT_NE(config_str.find(string), std::string::npos) << "'" << string << "' not found";
   }
@@ -182,6 +182,29 @@ TEST(TestConfig, RemainingTemplatesThrows) {
   } catch (std::runtime_error& err) {
     EXPECT_EQ(err.what(), std::string("could not resolve all template keys in config"));
   }
+}
+
+TEST(TestConfig, EnablePlatformCertificatesValidation) {
+  auto engine_builder = EngineBuilder();
+  envoy::config::bootstrap::v3::Bootstrap bootstrap;
+  engine_builder.enablePlatformCertificatesValidation(false);
+  auto config_str1 = engine_builder.generateConfigStr();
+  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str1), bootstrap);
+  ASSERT_THAT(bootstrap.DebugString(),
+              Not(HasSubstr("envoy_mobile.cert_validator.platform_bridge_cert_validator")));
+  ASSERT_THAT(bootstrap.DebugString(), HasSubstr("trusted_ca"));
+
+#if not defined(__APPLE__)
+  engine_builder.enablePlatformCertificatesValidation(true);
+  auto config_str2 = engine_builder.generateConfigStr();
+  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str2), bootstrap);
+  ASSERT_THAT(bootstrap.DebugString(),
+              HasSubstr("envoy_mobile.cert_validator.platform_bridge_cert_validator"));
+  ASSERT_THAT(bootstrap.DebugString(), Not(HasSubstr("trusted_ca")));
+#else
+  EXPECT_DEATH(engine_builder.enablePlatformCertificatesValidation(true),
+               "Certificates validation using platform provided APIs is not supported in IOS");
+#endif
 }
 
 } // namespace